Split rules and actual a11y rule implementation
